Policy operations are moved to test_user_policy_crud.py (#1685)
* Policy operations are moved to test_user_policy_crud.py


* Policy operations are moved to test_user_policy_crud.py


* Policy operations are moved to test_user_policy_crud.py
