+++ requirements.txt
@@ -71,3 +71,4 @@ urllib3==1.26.5
 Werkzeug==1.0.1
 word2number~=1.1
 xmltodict==0.12.0
+asyncio==3.4.3
